feat: Security, Exception, DTO

+++ src/main/java/net/codejava/sprint_boot_cadastrar_usuario/service/UsuarioService.java
@@ -1,5 +1,7 @@
 package net.codejava.sprint_boot_cadastrar_usuario.service;
 
+import net.codejava.sprint_boot_cadastrar_usuario.dto.UsuarioDTO;
+import net.codejava.sprint_boot_cadastrar_usuario.exception.ResourceNotFoundException;
 import net.codejava.sprint_boot_cadastrar_usuario.model.Usuario;
 import net.codejava.sprint_boot_cadastrar_usuario.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class UsuarioService {
@@ -14,20 +17,52 @@ public class UsuarioService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
-    public Usuario inserirUsuario(Usuario usuario) {
-        return usuarioRepository.save(usuario);
+    public UsuarioDTO inserirUsuario(UsuarioDTO usuarioDTO) {
+        Usuario usuario = new Usuario(usuarioDTO.getNome(), usuarioDTO.getEmail(), usuarioDTO.getSenha());
+        Usuario usuarioInserido = usuarioRepository.save(usuario);
+        return new UsuarioDTO(usuarioInserido.getId(), usuarioInserido.getNome(), usuarioInserido.getEmail()
+                , usuarioInserido.getSenha());
     }
 
-    public List<Usuario> buscarTodosUsuarios() {
-        return usuarioRepository.findAll();
+    public List<UsuarioDTO> buscarTodosUsuarios() {
+        List<Usuario> usuarios = usuarioRepository.findAll();
+        return usuarios.stream()
+                .map(usuario -> new UsuarioDTO(usuario.getId(), usuario.getNome(), usuario.getEmail()
+                        , usuario.getSenha()))
+                .collect(Collectors.toList());
     }
 
-    public Optional<Usuario> buscarUsuarioId(Long id) {
-        return usuarioRepository.findById(id);
+    public Optional<UsuarioDTO> buscarUsuarioId(Long id) {
+        return usuarioRepository.findById(id)
+                .map(usuario -> new UsuarioDTO(usuario.getId(), usuario.getNome(), usuario.getEmail()
+                        , usuario.getSenha()));
     }
 
     public void deletarUsuario(Long id) {
+        if (!usuarioRepository.existsById(id)) {
+            throw  new ResourceNotFoundException("Usuario com o ID " + id + " não encontrado");
+        }
         usuarioRepository.deleteById(id);
+    }
+
+    public UsuarioDTO atualizarUsuario(Long id, UsuarioDTO usuarioDTO) {
+        Usuario usuarioExistente = usuarioRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario com o ID " + id + " não encontrado"));
+
+        if (usuarioDTO.getNome() != null) {
+            usuarioExistente.setNome(usuarioDTO.getNome());
+        }
+        if (usuarioDTO.getEmail() != null) {
+            usuarioExistente.setEmail(usuarioDTO.getEmail());
+        }
+        if (usuarioDTO.getSenha() != null) {
+            usuarioExistente.setSenha(usuarioDTO.getSenha());
+        }
+
+        Usuario usuarioAtualizado = usuarioRepository.save(usuarioExistente);
+
+        return new UsuarioDTO(usuarioAtualizado.getId(), usuarioAtualizado.getNome(), usuarioAtualizado.getEmail()
+                , usuarioAtualizado.getSenha());
     }
 
 }
